Implemented better security and api calls to check login status and logout

+++ src/services/postCleanupService.ts
@@ -2,8 +2,14 @@ import cron from 'node-cron';
 import { Op } from 'sequelize';
 import Post from '../models/post';
 
-const startPostCleanupJob = () => {
-    cron.schedule('*/5 * * * *', async () => {
+let cleanupTask: cron.ScheduledTask | null = null;
+let isPostCleanupRunning: boolean = false;
+
+export const startPostCleanupJob = () => {
+    if (isPostCleanupRunning) return;
+    isPostCleanupRunning = true;
+
+    cleanupTask = cron.schedule('*/5 * * * *', async () => {
         try {
             const twoHoursAgo = new Date(Date.now() - 2 * 60 * 60 * 1000);
             const deletedPosts = await Post.destroy({
@@ -20,4 +26,10 @@ const startPostCleanupJob = () => {
     })
 }
 
-export default startPostCleanupJob;
+export const stopPostCleanupJob = () => {
+    if (cleanupTask) {
+        cleanupTask.stop();
+        console.log('Cron job stopped.');
+    }
+}
+
